Adding User Functinality

+++ controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 // @desc Set goal
 // @route POST /api/goal
 // @access Private
@@ -10,7 +11,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error("Please Add a Text Field")
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -19,7 +21,7 @@ const setGoal = asyncHandler(async (req, res) => {
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -32,6 +34,15 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal Not Found')
     }
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(400)
+        throw new Error('User Not Found')
+    }
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User Not Authorize')
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
     res.status(200).json(updatedGoal)
 
@@ -43,8 +54,17 @@ const updateGoal = asyncHandler(async (req, res) => {
 const deleteGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id)
     if (!goal) {
-        res.status(4000)
+        res.status(400)
         throw new Error('Goal Not Found')
+    }
+    const user = await User.findById(req.user.id)
+    if (!user) {
+        res.status(400)
+        throw new Error('User Not Found')
+    }
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User Not Authorize')
     }
     // await Goal.findByIdAndDelete(req.params.id)
     await Goal.findByIdAndRemove(req.params.id)
